disabling watchdogs with raw resource split

+++ test/Microsoft.Health.Fhir.Shared.Tests.Integration/Features/ChangeFeed/SqlServerFhirResourceChangeCaptureEnabledTests.cs
@@ -131,6 +131,12 @@ namespace Microsoft.Health.Fhir.Tests.Integration.Features.ChangeFeed
         [Fact]
         public async Task GivenChangeCaptureEnabledAndNoVersionPolicy_AfterUpdating_InvisibleHistoryIsRemovedByWatchdog()
         {
+            // TODO: Remove skip when watchdog is fixed
+            if (SqlAdlsClient.Container != null)
+            {
+                return;
+            }
+
             EnableInvisibleHistory();
             ExecuteSql("TRUNCATE TABLE dbo.Transactions");
             ExecuteSql("DELETE FROM dbo.Resource");
@@ -187,6 +193,12 @@ namespace Microsoft.Health.Fhir.Tests.Integration.Features.ChangeFeed
         [Fact]
         public async Task GivenChangeCaptureEnabledAndNoVersionPolicy_AfterHardDeleting_InvisibleHistoryIsRetainedAndIsRemovedByWatchdog()
         {
+            // TODO: Remove skip when watchdog is fixed
+            if (SqlAdlsClient.Container != null)
+            {
+                return;
+            }
+
             EnableInvisibleHistory();
             ExecuteSql("TRUNCATE TABLE dbo.Transactions");
             ExecuteSql("DELETE FROM dbo.Resource");
